# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChallengeCoverageResponse": {
        "properties": {
          "covered": {
            "items": {
              "enum": [
                "unsupported_assumption",
                "contradictory_evidence",
                "hidden_dependency",
                "failure_mode",
                "causal_claim",
                "missing_success_criteria"
              ],
              "type": "string"
            },
            "title": "Covered",
            "type": "array"
          },
          "uncovered": {
            "items": {
              "enum": [
                "unsupported_assumption",
                "contradictory_evidence",
                "hidden_dependency",
                "failure_mode",
                "causal_claim",
                "missing_success_criteria"
              ],
              "type": "string"
            },
            "title": "Uncovered",
            "type": "array"
          }
        },
        "required": [
          "covered",
          "uncovered"
        ],
        "title": "ChallengeCoverageResponse",
        "type": "object"
      },
      "ChallengeFindingCreate": {
        "properties": {
          "contribution_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution Id"
          },
          "detail": {
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "kind": {
            "enum": [
              "unsupported_assumption",
              "contradictory_evidence",
              "hidden_dependency",
              "failure_mode",
              "causal_claim",
              "missing_success_criteria"
            ],
            "title": "Kind",
            "type": "string"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "severity",
          "detail"
        ],
        "title": "ChallengeFindingCreate",
        "type": "object"
      },
      "ChallengeFindingResponse": {
        "properties": {
          "contribution_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "unsupported_assumption",
              "contradictory_evidence",
              "hidden_dependency",
              "failure_mode",
              "causal_claim",
              "missing_success_criteria"
            ],
            "title": "Kind",
            "type": "string"
          },
          "lang": {
            "enum": [
              "fr",
              "en"
            ],
            "title": "Lang",
            "type": "string"
          },
          "origin": {
            "enum": [
              "human",
              "critic"
            ],
            "title": "Origin",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "enum": [
              "proposed",
              "confirmed",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "kind",
          "severity",
          "detail",
          "origin",
          "status",
          "lang",
          "created_at",
          "updated_at"
        ],
        "title": "ChallengeFindingResponse",
        "type": "object"
      },
      "ChallengeListResponse": {
        "properties": {
          "coverage": {
            "$ref": "#/components/schemas/ChallengeCoverageResponse"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ChallengeFindingResponse"
            },
            "title": "Findings",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/ChallengeRunResponse"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "findings",
          "coverage"
        ],
        "title": "ChallengeListResponse",
        "type": "object"
      },
      "ChallengeRunDetailResponse": {
        "properties": {
          "coverage": {
            "$ref": "#/components/schemas/ChallengeCoverageResponse"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ChallengeFindingResponse"
            },
            "title": "Findings",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/ChallengeRunResponse"
          }
        },
        "required": [
          "run",
          "findings",
          "coverage"
        ],
        "title": "ChallengeRunDetailResponse",
        "type": "object"
      },
      "ChallengeRunResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lang": {
            "enum": [
              "fr",
              "en"
            ],
            "title": "Lang",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "opened_by": {
            "format": "uuid",
            "title": "Opened By",
            "type": "string"
          },
          "option_id": {
            "format": "uuid",
            "title": "Option Id",
            "type": "string"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "space_id",
          "option_id",
          "status",
          "opened_by",
          "lang",
          "created_at",
          "updated_at"
        ],
        "title": "ChallengeRunResponse",
        "type": "object"
      },
...
      "FindingResolutionRequest": {
        "properties": {
          "resolution": {
            "enum": [
              "confirmed",
              "dismissed"
            ],
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "resolution"
        ],
        "title": "FindingResolutionRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "member"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Kollio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/workspaces": {
      "get": {
        "operationId": "list_workspaces",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceResponse"
                  },
                  "title": "Response List Workspaces",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workspaces",
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/workspaces/{workspace_id}/decision-spaces/{space_id}/options/{option_id}/challenges": {
      "get": {
        "operationId": "list_challenges",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Option Challenges",
        "tags": [
          "challenge"
        ]
      },
      "post": {
        "operationId": "open_challenge",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Open Option Challenge",
        "tags": [
          "challenge"
        ]
      }
    },
    "/workspaces/{workspace_id}/decision-spaces/{space_id}/options/{option_id}/challenges/{run_id}": {
      "get": {
        "operationId": "get_challenge",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Option Challenge",
        "tags": [
          "challenge"
        ]
      }
    },
    "/workspaces/{workspace_id}/decision-spaces/{space_id}/options/{option_id}/challenges/{run_id}/completion": {
      "post": {
        "operationId": "complete_challenge",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Option Challenge",
        "tags": [
          "challenge"
        ]
      }
    },
    "/workspaces/{workspace_id}/decision-spaces/{space_id}/options/{option_id}/challenges/{run_id}/findings": {
      "post": {
        "operationId": "record_challenge_finding",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeFindingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeFindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Challenge Finding",
        "tags": [
          "challenge"
        ]
      }
    },
    "/workspaces/{workspace_id}/decision-spaces/{space_id}/options/{option_id}/challenges/{run_id}/findings/{finding_id}/resolution": {
      "post": {
        "operationId": "resolve_challenge_finding",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeFindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Challenge Finding",
        "tags": [
          "challenge"
        ]
      }
    },
...
  }
}